Split up backend config file to be more modular - Now there are new files in the config directory, 'express-middleware', 'apollo-server', and 'mongo' that each export a function that initialize their respective parts of the server.
