+++ backend/graphql/schema/index.js
@@ -4,10 +4,4 @@ const Team = require('./Team');
 const MsgComment = require('./MsgComment');
 const User = require('./User');
 
-module.exports = {
-	Tag,
-	Message,
-	Team,
-	MsgComment,
-	User
-};
+module.exports = [Tag, Message, Team, MsgComment, User];
